Conflicts:
	ground/openpilotgcs/src/plugins/config/configgadgetwidget.cpp

+++ flight/PiOS/inc/pios_usb.h
@@ -36,7 +36,7 @@
 extern int32_t PIOS_USB_Reenumerate();
 extern int32_t PIOS_USB_ChangeConnectionState(bool connected);
 extern bool PIOS_USB_CableConnected(uint8_t id);
-extern bool PIOS_USB_CheckAvailable(uint8_t id);
+extern bool PIOS_USB_CheckAvailable(uint32_t id);
 
 #endif /* PIOS_USB_H */
